+++ docs/ChangeLog.md
@@ -106,6 +106,7 @@
 - edit: revert edits with widget `discard-line` (reported by dezza) `#D2301` 3b2b4b81
 - vi_nmap: fix cursor position after <kbd>C-o</kbd> `#D2302` c106239a
 - decode (`ble-bind`): initialize specified keymaps (motivated by quantumfrost) `#D2324` 66e450d7
+- edit (`display-shell-version`): show `(integration: off)` for plugins with integration turned off `#D2330` xxxxxxxx
 
 ## Fixes